+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/mixin/legal/LicenseMixIn.java
@@ -2,8 +2,9 @@ package de.digitalcollections.model.jackson.mixin.legal;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import de.digitalcollections.model.jackson.mixin.UniqueObjectMixIn;
 import de.digitalcollections.model.legal.License;
 
 @JsonDeserialize(as = License.class)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public interface LicenseMixIn {}
+public interface LicenseMixIn extends UniqueObjectMixIn {}