+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/mixin/security/UserMixIn.java
@@ -1,11 +1,10 @@
 package de.digitalcollections.model.jackson.mixin.security;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import de.digitalcollections.model.jackson.mixin.UniqueObjectMixIn;
 import de.digitalcollections.model.security.User;
 
 @JsonDeserialize(as = User.class)
 @JsonIgnoreProperties(ignoreUnknown = true)
-@JsonTypeName("USER")
-public interface UserMixIn {}
+public interface UserMixIn extends UniqueObjectMixIn {}